Tidy up floating-point precision

+++ src/sfac.h
@@ -24,11 +24,11 @@ struct mol_species
 	char species[4];    /* Species name */
 	int n_atoms;        /* Number of atoms of this species */
 
-	float x[32*1024];
-	float y[32*1024];
-	float z[32*1024];
-	float occ[32*1024];
-	float B[32*1024];
+	double x[32*1024];
+	double y[32*1024];
+	double z[32*1024];
+	double occ[32*1024];
+	double B[32*1024];
 };
